fixing-billing-Initial-Request
Matthew is launching several new Allocations based ticket types. In order for him to better organize those, he wants to change the code for a billing initial Request (the ticket created when a Quick Ticket - Billing is created)
This change is only live in Sandbox. The old code is:
<option data-id="issue-4" data-support-tag="ALR4" data-value="Questions About My Monthly Statement" value="qamms">
and
<option data-id="issue-15" data-support-tag="ALR4" data-value="Other Billing Issues" value="obi">
So we need to change ALR4 with BIL1. So we need to update the staging/sandbox enviroment