Create automated test with Gauge
This is needed test for the clean-web front end. I have built it to be associated with Typescript and was using Taiko as a core to express the test.
Other things that comming along with this commit are the scenarios for ticket #444. Currently, I solved the below scenarios:
	- Scenario 2: Verify Global Admin can access to create equipment page
	- Scenario 4: Verify Validations are appeared incase user hasn’t finished with the form information
	- Scenario 6: Verify User can create a device successfully
But there is still some small issues for scenario 6 because currently, on localhost does not display device list to verify this step. But it will be fixed in later changes
For other scenarios, we have some blockers as known below:
	- Scenario 1: Verify Admin users at a company can access to create equipment page
Status: Block. There is no admin account to test
	- Scenario 3: Verify other user (not Admin user) can NOT access to create equipment page
Status: Block. There is no non-admin account to test
	- Scenario 5: Verify data binding are correct for each fields type
Status: Block. Currently, website does not support data binding
	- Scenario 7: Verify the device submitted by Admin users at a company is moved to that company’s inventory
Status: Block. There is no admin account to test
	- Scenario 8: Verify the device submitted by Global Admin users at a company is moved to the company’s inventory that he entered at Company section
Status: Block. Currently, user cannot switch to Other company's inventory